# kernel-openapi.yaml
openapi: 3.1.0
info:
  title: Institutional Kernel API
  version: "1.1.0"
  description: >
    API HTTP del Institutional Kernel de Mi Rotaract. Fuente única de verdad
    para identidad, organizaciones jerárquicas, membresías, períodos, cargos
    y autoridades, autorización contextual, solicitudes de ingreso,
    transferencias entre organizaciones y registro de módulos externos.
    Este documento es un contrato de datos derivado de `kernel-spec.md`
    (secciones 4, 5, 9 y 10). No reemplaza la especificación; ante cualquier
    conflicto, `kernel-spec.md` es la fuente normativa. El proceso de
    compatibilidad y publicación está definido en
    `kernel-openapi-contract-plan.md`.
  contact:
    name: Institutional Kernel
  license:
    name: Uso interno — Mi Rotaract
    identifier: LicenseRef-MiRotaract-Internal
...
servers:
  - url: https://api.agendai.com.ar/api/kernel/v1
    description: Producción
  - url: https://staging-api.agendai.com.ar/api/kernel/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /permissions:
    get:
      operationId: listPermissions
      tags: [Authorization]
      summary: Listar definiciones de permiso
      x-required-permission: kernel.role.read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PermissionDefinition" }
    post:
      operationId: registerPermission
      tags: [Authorization]
      summary: Registrar una definición de permiso
      description: >
        Un módulo sólo puede registrar permisos dentro de su namespace
        (invariante 6.7.9). No se agregan permisos mediante enums compilados
        (invariante 6.7.2).
      x-required-permission: kernel.role.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePermissionRequest" }
      responses:
        "201":
          description: Creado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PermissionDefinition" }
  /roles:
    get:
      operationId: listRoles
      tags: [Authorization]
      summary: Listar definiciones de rol
      x-required-permission: kernel.role.read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RoleDefinition" }
    post:
      operationId: createRole
      tags: [Authorization]
      summary: Crear una definición de rol
      x-required-permission: kernel.role.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRoleRequest" }
      responses:
        "201":
          description: Creado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleDefinition" }
  /roles/{roleId}/permissions/{permissionId}:
    put:
      operationId: attachPermissionToRole
      tags: [Authorization]
      summary: Adjuntar un permiso a un rol
      x-required-permission: kernel.role.manage
      parameters:
        - $ref: "#/components/parameters/roleId"
        - $ref: "#/components/parameters/permissionId"
      responses:
        "204": { description: Adjuntado. }
    delete:
      operationId: detachPermissionFromRole
      tags: [Authorization]
      summary: Quitar un permiso de un rol
      x-required-permission: kernel.role.manage
      parameters:
        - $ref: "#/components/parameters/roleId"
        - $ref: "#/components/parameters/permissionId"
      responses:
        "204": { description: Quitado. }
  /role-assignments:
    post:
      operationId: grantRole
      tags: [Authorization]
      summary: Otorgar (o denegar explícitamente) un rol en un alcance
      description: >
        `PLATFORM` sólo puede asignarse a roles expresamente autorizados
        (invariante 6.7.7).
      x-required-permission: kernel.role.assign
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GrantRoleRequest" }
      responses:
        "201":
          description: Creado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleAssignment" }
    get:
      operationId: listRoleAssignments
      tags: [Authorization]
      summary: Listar asignaciones de rol
      x-required-permission: kernel.role.read
      parameters:
        - name: personId
          in: query
          schema: { type: string }
        - name: organizationId
          in: query
          schema: { type: string }
        - name: roleDefinitionId
          in: query
          schema: { type: string }
        - name: includeRevoked
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RoleAssignment" }
  /role-assignments/{assignmentId}/revoke:
    post:
      operationId: revokeRole
      tags: [Authorization]
      summary: Revocar una asignación de rol
      x-required-permission: kernel.role.revoke
      parameters:
        - $ref: "#/components/parameters/assignmentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleAssignment" }
  /authorization/check:
    post:
      operationId: checkAuthorization
      tags: [Authorization]
      summary: Evaluar una decisión de autorización puntual
      description: >
        Denegación por defecto (principio 9). `DENY` gana sobre `ALLOW` con
        igual o mayor especificidad (invariante 6.7.4).
        Un sujeto autenticado siempre puede consultar su propia
        autorización (`subjectId` = persona propia); consultar la de otro
        `subjectId` requiere `kernel.role.read`.
      x-required-permission: kernel.role.read
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthorizationCheckRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthorizationDecision" }
  /authorization/batch-check:
    post:
      operationId: batchCheckAuthorization
      tags: [Authorization]
      summary: Evaluar hasta 100 decisiones de autorización (CA-AUTHZ-06)
      x-required-permission: kernel.role.read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [checks]
              properties:
                checks:
                  type: array
                  maxItems: 100
                  items:
                    { $ref: "#/components/schemas/AuthorizationCheckRequest" }
      responses:
        "200":
          description: >
            Devuelve una decisión por cada ítem, en el mismo orden
            (CA-AUTHZ-06).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuthorizationDecision" }
  /persons/{personId}/effective-permissions:
    get:
      operationId: getEffectivePermissions
      tags: [Authorization]
      summary: Listar permisos efectivos de una persona en un alcance
      x-required-permission: kernel.role.read
      parameters:
        - $ref: "#/components/parameters/personId"
        - name: organizationId
          in: query
          schema: { type: string }
        - name: periodId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
              example: ["kernel.membership.read", "meetings.meeting.create"]

  # ---------------------------------------------------------------------
  # 9.9 Solicitudes
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token de usuario emitido por `POST /auth/login` o
        `/auth/refresh` (§14.2). Vida útil de 10 minutos. `aud`:
        `mirotaract-platform`.
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token de servicio a servicio firmado (o mTLS equivalente, §14.3).
        `sub` identifica al servicio emisor. `aud` debe ser
        `institutional-kernel`. Requerido para todos los endpoints bajo
        `/service/*`; el kernel rechaza estas rutas con token de usuario.
  parameters:
    personId:
      name: personId
      in: path
      required: true
      schema: { type: string }
...
    roleId:
      name: roleId
      in: path
      required: true
      schema: { type: string }
    permissionId:
      name: permissionId
      in: path
      required: true
      schema: { type: string }
    assignmentId:
      name: assignmentId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    ScopeType:
      type: string
      enum: [PLATFORM, ORGANIZATION, ORGANIZATION_TREE]
    AssignmentEffect:
      type: string
      enum: [ALLOW, DENY]
...
    PermissionDefinition:
      type: object
      required: [id, code, namespace, name, isSystem, createdAt, updatedAt]
      properties:
        id: { type: string }
        code:
          type: string
          description: Formato `<namespace>.<resource>.<action>` (invariante 6.7.1).
        namespace: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        resourceType: { type: string, nullable: true }
        moduleId: { type: string, nullable: true }
        isSystem: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreatePermissionRequest:
      type: object
      required: [code, namespace, name]
      properties:
        code: { type: string }
        namespace: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        resourceType: { type: string, nullable: true }
        moduleId: { type: string, nullable: true }
    RoleDefinition:
      type: object
      required: [id, code, name, isSystem, createdAt, updatedAt]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        moduleId: { type: string, nullable: true }
        isSystem: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateRoleRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        moduleId: { type: string, nullable: true }
    RoleAssignment:
      type: object
      required:
        [
          id,
          personId,
          roleDefinitionId,
          effect,
          scopeType,
          validFrom,
          createdAt,
        ]
      properties:
        id: { type: string }
        personId: { type: string }
        roleDefinitionId: { type: string }
        effect: { $ref: "#/components/schemas/AssignmentEffect" }
        scopeType: { $ref: "#/components/schemas/ScopeType" }
        organizationId:
          type: string
          nullable: true
          description: Requerido si `scopeType` != PLATFORM (invariante de esquema §5.1).
        periodId: { type: string, nullable: true }
        validFrom: { type: string, format: date-time }
        validUntil: { type: string, format: date-time, nullable: true }
        grantedById: { type: string, nullable: true }
        reason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        revokedAt: { type: string, format: date-time, nullable: true }
        revokedById: { type: string, nullable: true }
        sourceAppointmentId:
          type: string
          nullable: true
          readOnly: true
          description: FK obligatoria para asignaciones derivadas de un cargo; null si fue asignada manualmente.
    GrantRoleRequest:
      type: object
      required: [personId, roleDefinitionId, scopeType]
      properties:
        personId: { type: string }
        roleDefinitionId: { type: string }
        effect: { $ref: "#/components/schemas/AssignmentEffect" }
        scopeType: { $ref: "#/components/schemas/ScopeType" }
        organizationId: { type: string, nullable: true }
        periodId: { type: string, nullable: true }
        validFrom: { type: string, format: date-time }
        validUntil: { type: string, format: date-time, nullable: true }
        reason: { type: string, nullable: true }
    AuthorizationCheckRequest:
      type: object
      required: [subjectId, permission, scope]
      properties:
        subjectId:
          type: string
          description: personId a evaluar.
        permission:
          type: string
          example: meetings.meeting.create
        scope:
          type: object
          required: [type]
          properties:
            type: { $ref: "#/components/schemas/ScopeType" }
            organizationId: { type: string, nullable: true }
        periodId: { type: string, nullable: true }
        resource:
          type: object
          nullable: true
          properties:
            type: { type: string }
            id: { type: string }
            attributes: { type: object }
    AuthorizationDecision:
      type: object
      required: [allowed, decisionId, subjectId, permission, evaluatedAt]
      properties:
        allowed: { type: boolean }
        decisionId: { type: string }
        subjectId: { type: string }
        permission: { type: string }
        matchedAssignments:
          type: array
          items: { type: string }
        reasonCodes:
          type: array
          items: { type: string }
          example: ["ROLE_ALLOWED"]
        evaluatedAt: { type: string, format: date-time }
        cacheUntil: { type: string, format: date-time, nullable: true }
...